Safari-609.1.11.1/LayoutTests/fast/dom/HTMLButtonElement/click.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests HTMLButtonElement's click is just HTMLElement click");
shouldBe("HTMLElement.prototype.click", "HTMLButtonElement.prototype.click");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
