LayoutTests/fast/dom/HTMLButtonElement/click-expected.txt

Tests HTMLButtonElement's click is just HTMLElement click

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS HTMLElement.prototype.click is HTMLButtonElement.prototype.click
PASS successfullyParsed is true

TEST COMPLETE
